major rewrite of the library
