+++ Cargo.toml
@@ -16,6 +16,9 @@ codecov = {repository = "sile/hls_m3u8"}
 
 [dependencies]
 trackable = "0.2"
+derive_builder = "0.7.2"
+shrinkwraprs = "0.2.1"
+regex = { version = "1.2.1", features = [ "pattern" ] }
 
 [dev-dependencies]
 clap = "2"
