+++ rustfmt.toml
@@ -1,0 +1,2 @@
+error_on_line_overflow = true
+error_on_unformatted = true
