+++ src/error.rs
@@ -5,9 +5,11 @@ use trackable::error::{ErrorKind as TrackableErrorKind, TrackableError};
 pub struct Error(TrackableError<ErrorKind>);
 
 /// Possible error kinds.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
 #[allow(missing_docs)]
 pub enum ErrorKind {
     InvalidInput,
+    BuilderError(String),
 }
+
 impl TrackableErrorKind for ErrorKind {}
