+++ src/tags/iframe_stream_inf.rs
@@ -1,0 +1,179 @@
+use std::borrow::Cow;
+use std::fmt;
+use std::str::FromStr;
+
+use crate::attribute::AttributePairs;
+use crate::error::{Error, ErrorKind};
+use crate::types::{DecimalResolution, HdcpLevel, ProtocolVersion};
+use crate::utils::parse_u64;
+use crate::utils::{quote, unquote};
+
+/// [4.3.4.3. EXT-X-I-FRAME-STREAM-INF]
+///
+/// [4.3.4.3. EXT-X-I-FRAME-STREAM-INF]: https://tools.ietf.org/html/rfc8216#section-4.3.4.3
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+pub struct ExtXIFrameStreamInf {
+    uri: String,
+    bandwidth: u64,
+    average_bandwidth: Option<u64>,
+    codecs: Option<String>,
+    resolution: Option<DecimalResolution>,
+    hdcp_level: Option<HdcpLevel>,
+    video: Option<String>,
+}
+
+impl ExtXIFrameStreamInf {
+    pub(crate) const PREFIX: &'static str = "#EXT-X-I-FRAME-STREAM-INF:";
+
+    /// Makes a new `ExtXIFrameStreamInf` tag.
+    pub fn new<T: ToString>(uri: T, bandwidth: u64) -> Self {
+        ExtXIFrameStreamInf {
+            uri: uri.to_string(),
+            bandwidth,
+            average_bandwidth: None,
+            codecs: None,
+            resolution: None,
+            hdcp_level: None,
+            video: None,
+        }
+    }
+
+    /// Returns the URI that identifies the associated media playlist.
+    pub fn uri(&self) -> Cow<'_, str> {
+        Cow::Borrowed(&self.uri)
+    }
+
+    /// Returns the peak segment bit rate of the variant stream.
+    pub fn bandwidth(&self) -> u64 {
+        self.bandwidth
+    }
+
+    /// Returns the average segment bit rate of the variant stream.
+    pub fn average_bandwidth(&self) -> Option<u64> {
+        self.average_bandwidth
+    }
+
+    /// Returns a string that represents the list of codec types contained the variant stream.
+    pub fn codecs(&self) -> Option<Cow<'_, str>> {
+        match &self.codecs {
+            Some(value) => Some(Cow::Borrowed(&value)),
+            None => None,
+        }
+    }
+
+    /// Returns the optimal pixel resolution at which to display all the video in the variant stream.
+    pub fn resolution(&self) -> Option<(usize, usize)> {
+        match self.resolution {
+            Some(value) => Some((value.width(), value.height())),
+            None => None,
+        }
+    }
+
+    /// Returns the HDCP level of the variant stream.
+    pub fn hdcp_level(&self) -> Option<HdcpLevel> {
+        self.hdcp_level
+    }
+
+    /// Returns the group identifier for the video in the variant stream.
+    pub fn video(&self) -> Option<Cow<'_, str>> {
+        match &self.video {
+            Some(value) => Some(Cow::Borrowed(&value)),
+            None => None,
+        }
+    }
+
+    /// Returns the protocol compatibility version that this tag requires.
+    pub const fn required_version(&self) -> ProtocolVersion {
+        ProtocolVersion::V1
+    }
+}
+
+impl fmt::Display for ExtXIFrameStreamInf {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "{}", Self::PREFIX)?;
+        write!(f, "URI={}", quote(&self.uri))?;
+        write!(f, ",BANDWIDTH={}", self.bandwidth)?;
+
+        if let Some(value) = &self.average_bandwidth {
+            write!(f, ",AVERAGE-BANDWIDTH={}", value)?;
+        }
+
+        if let Some(value) = &self.codecs {
+            write!(f, ",CODECS={}", quote(value))?;
+        }
+
+        if let Some(value) = &self.resolution {
+            write!(f, ",RESOLUTION={}", value)?;
+        }
+
+        if let Some(value) = &self.hdcp_level {
+            write!(f, ",HDCP-LEVEL={}", value)?;
+        }
+
+        if let Some(value) = &self.video {
+            write!(f, ",VIDEO={}", quote(value))?;
+        }
+
+        Ok(())
+    }
+}
+
+impl FromStr for ExtXIFrameStreamInf {
+    type Err = Error;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        track_assert!(s.starts_with(Self::PREFIX), ErrorKind::InvalidInput);
+
+        let mut uri = None;
+        let mut bandwidth = None;
+        let mut average_bandwidth = None;
+        let mut codecs = None;
+        let mut resolution = None;
+        let mut hdcp_level = None;
+        let mut video = None;
+
+        let attrs = (s.split_at(Self::PREFIX.len()).1).parse::<AttributePairs>()?;
+
+        for (key, value) in attrs {
+            match key.as_str() {
+                "URI" => uri = Some(unquote(value)),
+                "BANDWIDTH" => bandwidth = Some(track!(parse_u64(value))?),
+                "AVERAGE-BANDWIDTH" => average_bandwidth = Some(track!(parse_u64(value))?),
+                "CODECS" => codecs = Some(unquote(value)),
+                "RESOLUTION" => resolution = Some(track!(value.parse())?),
+                "HDCP-LEVEL" => hdcp_level = Some(track!(value.parse())?),
+                "VIDEO" => video = Some(unquote(value)),
+                _ => {
+                    // [6.3.1. General Client Responsibilities]
+                    // > ignore any attribute/value pair with an unrecognized AttributeName.
+                }
+            }
+        }
+
+        let uri = track_assert_some!(uri, ErrorKind::InvalidInput);
+        let bandwidth = track_assert_some!(bandwidth, ErrorKind::InvalidInput);
+        Ok(ExtXIFrameStreamInf {
+            uri,
+            bandwidth,
+            average_bandwidth,
+            codecs,
+            resolution,
+            hdcp_level,
+            video,
+        })
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn ext_x_i_frame_stream_inf() {
+        let tag = ExtXIFrameStreamInf::new("foo".to_string(), 1000);
+        let text = r#"#EXT-X-I-FRAME-STREAM-INF:URI="foo",BANDWIDTH=1000"#;
+        assert_eq!(text.parse().ok(), Some(tag.clone()));
+        assert_eq!(tag.to_string(), text);
+        assert_eq!(tag.required_version(), ProtocolVersion::V1);
+    }
+}
