+++ src/tags/start.rs
@@ -1,35 +1,10 @@
-use super::parse_yes_or_no;
-use crate::attribute::AttributePairs;
-use crate::types::{ProtocolVersion, SignedDecimalFloatingPoint};
-use crate::{Error, ErrorKind, Result};
 use std::fmt;
 use std::str::FromStr;
 
-/// [4.3.5.1. EXT-X-INDEPENDENT-SEGMENTS]
-///
-/// [4.3.5.1. EXT-X-INDEPENDENT-SEGMENTS]: https://tools.ietf.org/html/rfc8216#section-4.3.5.1
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct ExtXIndependentSegments;
-impl ExtXIndependentSegments {
-    pub(crate) const PREFIX: &'static str = "#EXT-X-INDEPENDENT-SEGMENTS";
-
-    /// Returns the protocol compatibility version that this tag requires.
-    pub fn requires_version(self) -> ProtocolVersion {
-        ProtocolVersion::V1
-    }
-}
-impl fmt::Display for ExtXIndependentSegments {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        Self::PREFIX.fmt(f)
-    }
-}
-impl FromStr for ExtXIndependentSegments {
-    type Err = Error;
-    fn from_str(s: &str) -> Result<Self> {
-        track_assert_eq!(s, Self::PREFIX, ErrorKind::InvalidInput);
-        Ok(ExtXIndependentSegments)
-    }
-}
+use crate::attribute::AttributePairs;
+use crate::types::{ProtocolVersion, SignedDecimalFloatingPoint};
+use crate::utils::parse_yes_or_no;
+use crate::{Error, ErrorKind};
 
 /// [4.3.5.2. EXT-X-START]
 ///
@@ -39,28 +14,29 @@ pub struct ExtXStart {
     time_offset: SignedDecimalFloatingPoint,
     precise: bool,
 }
+
 impl ExtXStart {
     pub(crate) const PREFIX: &'static str = "#EXT-X-START:";
 
     /// Makes a new `ExtXStart` tag.
-    pub fn new(time_offset: SignedDecimalFloatingPoint) -> Self {
-        ExtXStart {
-            time_offset,
+    pub fn new(time_offset: f64) -> crate::Result<Self> {
+        Ok(Self {
+            time_offset: SignedDecimalFloatingPoint::new(time_offset)?,
             precise: false,
-        }
+        })
     }
 
     /// Makes a new `ExtXStart` tag with the given `precise` flag.
-    pub fn with_precise(time_offset: SignedDecimalFloatingPoint, precise: bool) -> Self {
-        ExtXStart {
-            time_offset,
+    pub fn with_precise(time_offset: f64, precise: bool) -> crate::Result<Self> {
+        Ok(Self {
+            time_offset: SignedDecimalFloatingPoint::new(time_offset)?,
             precise,
-        }
+        })
     }
 
     /// Returns the time offset of the media segments in the playlist.
-    pub fn time_offset(&self) -> SignedDecimalFloatingPoint {
-        self.time_offset
+    pub fn time_offset(&self) -> f64 {
+        self.time_offset.as_f64()
     }
 
     /// Returns whether clients should not render media stream whose presentation times are
@@ -70,10 +46,11 @@ impl ExtXStart {
     }
 
     /// Returns the protocol compatibility version that this tag requires.
-    pub fn requires_version(&self) -> ProtocolVersion {
+    pub fn required_version(&self) -> ProtocolVersion {
         ProtocolVersion::V1
     }
 }
+
 impl fmt::Display for ExtXStart {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write!(f, "{}", Self::PREFIX)?;
@@ -84,17 +61,18 @@ impl fmt::Display for ExtXStart {
         Ok(())
     }
 }
+
 impl FromStr for ExtXStart {
     type Err = Error;
-    fn from_str(s: &str) -> Result<Self> {
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
         track_assert!(s.starts_with(Self::PREFIX), ErrorKind::InvalidInput);
 
         let mut time_offset = None;
         let mut precise = false;
-        let attrs = AttributePairs::parse(s.split_at(Self::PREFIX.len()).1);
-        for attr in attrs {
-            let (key, value) = track!(attr)?;
-            match key {
+        let attrs = (s.split_at(Self::PREFIX.len()).1).parse::<AttributePairs>()?;
+        for (key, value) in attrs {
+            match key.as_str() {
                 "TIME-OFFSET" => time_offset = Some(track!(value.parse())?),
                 "PRECISE" => precise = track!(parse_yes_or_no(value))?,
                 _ => {
@@ -117,26 +95,17 @@ mod test {
     use super::*;
 
     #[test]
-    fn ext_x_independent_segments() {
-        let tag = ExtXIndependentSegments;
-        let text = "#EXT-X-INDEPENDENT-SEGMENTS";
-        assert_eq!(text.parse().ok(), Some(tag));
-        assert_eq!(tag.to_string(), text);
-        assert_eq!(tag.requires_version(), ProtocolVersion::V1);
-    }
-
-    #[test]
     fn ext_x_start() {
-        let tag = ExtXStart::new(SignedDecimalFloatingPoint::new(-1.23).unwrap());
+        let tag = ExtXStart::new(-1.23).unwrap();
         let text = "#EXT-X-START:TIME-OFFSET=-1.23";
         assert_eq!(text.parse().ok(), Some(tag));
         assert_eq!(tag.to_string(), text);
-        assert_eq!(tag.requires_version(), ProtocolVersion::V1);
+        assert_eq!(tag.required_version(), ProtocolVersion::V1);
 
-        let tag = ExtXStart::with_precise(SignedDecimalFloatingPoint::new(1.23).unwrap(), true);
+        let tag = ExtXStart::with_precise(1.23, true).unwrap();
         let text = "#EXT-X-START:TIME-OFFSET=1.23,PRECISE=YES";
         assert_eq!(text.parse().ok(), Some(tag));
         assert_eq!(tag.to_string(), text);
-        assert_eq!(tag.requires_version(), ProtocolVersion::V1);
+        assert_eq!(tag.required_version(), ProtocolVersion::V1);
     }
 }
