+++ tests/f08e80da-bf1d-4e3d-8899-f0f6155f6efa.m3u8
@@ -1,0 +1,17 @@
+#EXTM3U
+#EXT-X-VERSION:5
+
+#EXT-X-MEDIA:TYPE=AUDIO,GROUP-ID="audio",NAME="English stereo",LANGUAGE="en",AUTOSELECT=YES,URI="f08e80da-bf1d-4e3d-8899-f0f6155f6efa_audio_1_stereo_128000.m3u8"
+
+#EXT-X-STREAM-INF:BANDWIDTH=628000,CODECS="avc1.42c00d,mp4a.40.2",RESOLUTION=320x180,AUDIO="audio"
+f08e80da-bf1d-4e3d-8899-f0f6155f6efa_video_180_250000.m3u8
+#EXT-X-STREAM-INF:BANDWIDTH=928000,CODECS="avc1.42c00d,mp4a.40.2",RESOLUTION=480x270,AUDIO="audio"
+f08e80da-bf1d-4e3d-8899-f0f6155f6efa_video_270_400000.m3u8
+#EXT-X-STREAM-INF:BANDWIDTH=1728000,CODECS="avc1.42c00d,mp4a.40.2",RESOLUTION=640x360,AUDIO="audio"
+f08e80da-bf1d-4e3d-8899-f0f6155f6efa_video_360_800000.m3u8
+#EXT-X-STREAM-INF:BANDWIDTH=2528000,CODECS="avc1.42c00d,mp4a.40.2",RESOLUTION=960x540,AUDIO="audio"
+f08e80da-bf1d-4e3d-8899-f0f6155f6efa_video_540_1200000.m3u8
+#EXT-X-STREAM-INF:BANDWIDTH=4928000,CODECS="avc1.42c00d,mp4a.40.2",RESOLUTION=1280x720,AUDIO="audio"
+f08e80da-bf1d-4e3d-8899-f0f6155f6efa_video_720_2400000.m3u8
+#EXT-X-STREAM-INF:BANDWIDTH=9728000,CODECS="avc1.42c00d,mp4a.40.2",RESOLUTION=1920x1080,AUDIO="audio"
+f08e80da-bf1d-4e3d-8899-f0f6155f6efa_video_1080_4800000.m3u8
